+++ docs/CHANGELOG.md
@@ -7,6 +7,13 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.0.8] - 2025-03-18
+
+### Fixed
+
+- Serialized errors for production environment.
+- Correct Github initialization on helper.
+
 ## [2.0.7] - 2025-03-18
 
 ### Added